cleans up some imports and makes names clearer, adds titleSlug to todayQuestion.json and dummy code for submit
