+++ questions/todayQuestion.json
@@ -1,1 +1,1 @@
-{"TODAYDATE": "2023-01-29", "LASTINDEX": 0, "QUESTIONID": 1509}
+{"TODAYDATE": "2023-02-02", "LASTINDEX": 3, "QUESTIONID": 1551, "QUESTIONNAME": "minimum-operations-to-make-array-equal"}
