+++ submit.py
@@ -1,0 +1,9 @@
+import json
+import discord
+import asyncio
+from discord.ext import commands
+from lastACs import getLastACs
+from lcUtils import *
+
+async def submit(ctx):
+    pass
